removed repeated paramter from typeOfSEssion
Each session in a conference should only have one type so it doesn't
really make sense for a session type to be repeatable. If a session were
to need to have two different types for some reason they can just be
split up as two different session within the same conference. That is
sort of the point of conferences having multiple sessions.